hevc: use ptrdiff_t for all strides

+++ libavcodec/hevcdsp.h
@@ -34,10 +34,10 @@ typedef struct HEVCDSPContext {
 
     void (*transform_add[4])(uint8_t *dst, int16_t *coeffs, ptrdiff_t _stride);
 
-    void (*sao_band_filter)(uint8_t *dst, uint8_t *src, int stride, int *sao_offset_val,
+    void (*sao_band_filter)(uint8_t *dst, uint8_t *src, ptrdiff_t stride, int *sao_offset_val,
                             int sao_left_class, int width, int height);
 
-    void (*sao_edge_filter)(uint8_t *dst, uint8_t *src, int stride, int *sao_offset_val,
+    void (*sao_edge_filter)(uint8_t *dst, uint8_t *src, ptrdiff_t stride, int *sao_offset_val,
                             int sao_eo_class, int at_top_border, int at_bottom_border,
                             int at_left_border, int at_right_border,
                             int width, int height);
